requirements/base.txt
cryptography==41.0.7
celery==5.3.6
django==5.1.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
django-filter==24.3
djangorestframework==3.15.2
django-modeltranslation==0.19.9
django-redis==5.4.0
django-phonenumber-field==8.0.0
flower==2.0.1
firebase-admin==6.5.0
mutagen==1.47.0
psycopg2-binary==2.9.9
python-environ==0.4.54
phonenumbers==8.13.46
pillow==10.4.0
redis==5.0.1